reuse cln rpc client in walletkit
