+++ src/server/modules/lightning/walletkit/lnwalletkit.service.ts
@@ -30,7 +30,7 @@ export class LightningWalletKitService implements OnModuleInit {
 
 	private initializeGrpcClients() {
 		if (this.type === 'lnd') this.grpc_client = this.lndService.initializeWalletKitClient();
-		if (this.type === 'cln') this.grpc_client = this.clnService.initializeLightningClient();
+		if (this.type === 'cln') this.grpc_client = this.clnService.initializeWalletKitClient();
 	}
 
 	private makeGrpcRequest(method: string, request: any): Promise<any> {
@@ -46,49 +46,8 @@ export class LightningWalletKitService implements OnModuleInit {
 		});
 	}
 
-	private mapClnAddresses(resp: any): LightningAddresses {
-		const entries: any[] = Array.isArray(resp?.addresses) ? resp.addresses : [];
-
-		const mkAddress = (addr: string) => ({
-			address: addr,
-			is_internal: 'false',
-			balance: 0,
-			derivation_path: '',
-			public_key: Buffer.alloc(0),
-		});
-
-		const bech32 = entries.map((e) => e?.bech32).filter((x: string) => !!x);
-		const p2tr = entries.map((e) => e?.p2tr).filter((x: string) => !!x);
-
-		const account_with_addresses: LightningAddresses['account_with_addresses'] = [];
-
-		if (bech32.length) {
-			account_with_addresses.push({
-				name: 'bech32',
-				address_type: LightningAddressType.WITNESS_PUBKEY_HASH,
-				derivation_path: '',
-				addresses: bech32.map(mkAddress),
-			});
-		}
-
-		if (p2tr.length) {
-			account_with_addresses.push({
-				name: 'p2tr',
-				address_type: LightningAddressType.TAPROOT_PUBKEY,
-				derivation_path: '',
-				addresses: p2tr.map(mkAddress),
-			});
-		}
-
-		return {account_with_addresses};
-	}
-
 	async getLightningAddresses(): Promise<LightningAddresses> {
-		// return this.makeGrpcRequest('ListAddresses', {});
 		if (this.type === 'lnd') return this.makeGrpcRequest('ListAddresses', {});
-		if (this.type === 'cln') {
-			const addresses = await this.makeGrpcRequest('ListAddresses', {});
-			return this.mapClnAddresses(addresses);
-		}
+		if (this.type === 'cln') return this.clnService.mapClnAddresses(await this.makeGrpcRequest('ListAddresses', {}));
 	}
 }
